% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculation_functions.R
\name{bray.2halves}
\alias{bray.2halves}
\title{Bray-Curtis similarity of two concatenated vectors}
\usage{
bray.2halves(x)
}
\arguments{
\item{x}{a vector composed of the two vectors concatenated together.}
}
\value{
Bray-Curtis similarity of the two concatenated vectors
}
\description{
A function to calculate Bray-Curtis dissimilarity index of two vectors that
 can be used in \code{apply} to return a vector of BCs for each row (used to calculate \code{simOE} in \code{oestats})
}
\examples{
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package="melbstreambiota", mustWork = TRUE),
                                        sheet = 3))
#Bray-Curtis similarity of first 2 sampprs
ab <- unlist(c(bugData[1,-1],bugData[2,-1]))
bray.2halves(ab)
}
\seealso{
\code{\link[vegan]{vegdist}}
}
