% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bugfams}
\alias{bugfams}
\title{"bestModelsBugfams"
SIGNAL and other index grades for 174 macroinvertebrate families}
\format{a data frame of 174 rows and 9 columns
\describe{
  \item{fam}{bugcode, consistent with EPA Victoria and AUSRIVAS conventions}
  \item{family}{family name}
  \item{SIGNAL2}{SIGNAL grade from Chessman 2003}
  \item{SIGNALWoV2001}{}
  \item{SIGNALWoV2003}{}
  \item{SIGNALWesPt2003}{}
  \item{SIGNALYarra2003}{}
  \item{Rheophily index}{from Bond et al 2012}
  \item{Comment}{}
  \item{note}{}
}}
\usage{
bugfams
}
\description{
A data frame listing macroinvertebrate families and their grades for various SIGNAL variants and Rheophily index
}
\keyword{datasets}
