% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prediction_functions.R
\name{bugModPred1}
\alias{bugModPred1}
\title{Predict occurrence of a macroinvertebrate family in a set of sample-pairs}
\usage{
bugModPred1(bugcode, sampprPreds)
}
\arguments{
\item{bugcode}{a 4-character code representing a macroivertebrate family:
one of the 59 \code{taxon.classes$fam}s}

\item{sampprPreds}{a data.frame of sample-pair codes (samppr) each with
associated environmental predictors.}
}
\value{
A list of two vectors with length = number of rows in \code{sampprPred},
with each row corresponding to the equivalent row in \code{sampprPred}.
\describe{
  \item{pred1}{predicted probabilities of occurrence under conditions specified in \code{sampprPred}}
  \item{pred1NHI}{predicted probabilities of occurrence under no human impact
  (AttImp_L9 set to zero, AttForest_L35W1000 set to 1)}
  }
}
\description{
Predict occurrence of a macroinvertebrate family in a set of sample-pairs
}
\details{
The sampprPreds table must contain the 10 predictor variables used in the
models: nspring,nriff,processN,AttImp_L9, AttForest_L35W1000,
CatchmentArea_km2_InclDams, meanAnnQ_mm,mnAnnAirTm_deg,
CatIgneous, and SRI_48mth_weighted. The function \code{collateBugSamppr} can
be used to compile such a table
}
\examples{
qt25MWstreamsCurrent <- bugModPred1("QT25",melbstreambiota::mwstreams)
# plotMWstreamsByVar(qt25MWstreamsCurrent$pred1,
#   style = "fixed", fixedBreaks = seq(0,1,0.2),
#   varName =
#    paste(melbstreambiota::taxon.classes$family[melbstreambiota::taxon.classes$fam == "QT25"],
#          "prob of occurrence"), legend.cex = 0.75)
}
\seealso{
\code{\link{collateBugSamppr}}, \code{\link{taxon.classes}}
}
