% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prediction_functions.R
\name{bugModPred59}
\alias{bugModPred59}
\title{Compile predictions of occurrence for all 59 macroinvertebrate families and resulting indices in a set of sample-pairs}
\usage{
bugModPred59(sampprPreds)
}
\arguments{
\item{sampprPreds}{a data.frame of sample-pair codes (samppr) each with
associated environmental predictors. If there is no field called samppr, then the field subc if used.}
}
\value{
A list containing:
two dataframes with number of rows = that of \code{sampprPreds},
(each row corresponding to the equivalent row in \code{sampprPred}) and 59
columns, corresponding to the 59 families listed in \code{taxon.classes}
\describe{
  \item{sampprPredsLumar}{ data frame with 4 columns, samppr, LUMaR, SIGNAL2, and number
  of sensitive families (nSensFams) resulting from the predicted probabilities of
  occurrence.  Number of rows = that of \code{sampprPred}, (each row corresponding
  to the equivalent row in \code{sampprPred})}
  \item{pred59}{data frame containing predicted probabilities of occurrence
  under conditions specified in \code{sampprPred}, with number of
  rows = that of \code{sampprPred}, (each row corresponding to the
  equivalent row in \code{sampprPred}) and 59 columns, corresponding
  to the 59 families listed in \code{taxon.classes}}
  }
}
\description{
Compile predictions of occurrence for all 59 macroinvertebrate families and resulting indices in a set of sample-pairs
}
\details{
The sampprPreds table must contain the 10 predictor variables used in the
models: nspring, nriff, processN, AttImp_L9, AttForest_L35W1000,
CatchmentArea_km2_InclDams, meanAnnQ_mm, mnAnnAirTm_deg,
CatIgneous, and SRI_48mth_weighted. The function \code{collateBugSamppr} can
be used to compile such a table.  Use \code{bugModPred1} to make equivalent predictions for a single family.
Note that the LUMaR and number of sensitive families produced by this function are PREDICTED values
given the environmental variables supplied in sampprPreds.  These values are likely to differ from observed
values that can be calculated using the lumar function from your own appropriately processed
macroinvertebrate data (see the melbstreambiota vignette for instructions).
to the obs.table argument in the \code{lumar} function.
}
\examples{
mwstreamsCurrent <- bugModPred59(melbstreambiota::mwstreams)
plotMWstreamsByVar(mwstreamsCurrent$sampprPredsLumar$lumar, nbreaks = 6,
                   style = "fixed", fixedBreaks = c(-1,-0.25,0,0.25,0.5,0.75,1),
                   varName = "LUMaR", legend.cex = 0.75)
}
\seealso{
\code{\link{collateBugSamppr}}, \code{\link{bugModPred1}}, \code{\link{taxon.classes}}
}
