% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculation_functions.R
\name{calcSIGNAL}
\alias{calcSIGNAL}
\title{Calculate SIGNAL and SIGNAL2 scores from macroinvertebrate data}
\usage{
calcSIGNAL(bugData)
}
\arguments{
\item{bugData}{Table with either A) 2 columns (\code{samppr}, \code{bugcode}), or A) one row per samppr,
with the first column \code{samppr} and the remaining column headings \code{bugcode}s}
}
\value{
A data.frame with 3 columns: samppr, matching sampprs$samppr; SIGNAL and SIGNAL2.
}
\description{
Collates macroinvertebrate data (in two possible formats) and returns SIGNAL and
SIGNAL2 scores for each sample.
}
\details{
While all other functions in this package require sample-pairs of rapid bioassessment
samples, this function makes no assumption. SIGNAL scores can be calculated for any sample method.
All bugcodes should be consistent with taxon.classes. See \url{http://www.mdfrc.org.au/bugguide/resources/AUSRIVAS_Taxacodes.pdf}
for a full list including taxa other than those in taxon.classes.  Note that the convention used in this package is to
omit the trailing '9999' from each code. If bugcodes in bugData include trailing 4 digits, these are removed
(this effectively combines any lower-taxonomic-level taxa to family-level.)
SIGNAL is calculated using all taxa listed in bugData assigned a score in
EPA Victoria (2003) Rapid bioassessment methodology for rivers and streams.
Report 604.1. Environment Protection Authority Victoria, Melbourne.  SIGNAL2 uses
the scores listed by Chessman, B.C. (2003) New sensitivity grades for Australian
river macroinvertebrates. Marine and Freshwater Research, 54, 95–103.
}
\examples{
#Data in Form A
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package="melbstreambiota", mustWork = TRUE),
                                        sheet = 2))
#When importing data, replace the system.file() statement with "path/excelFile.xls"
signalScores1 <- calcSIGNAL(bugData)
#Data in Form B
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package = "melbstreambiota", mustWork = TRUE),
                                        sheet = 3))
signalScores2 <- calcSIGNAL(bugData)
#Check that data imported in both forms are equql
sum(signalScores1[-1] - signalScores2[-1])  # = 0
}
