% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_functions.R
\name{collateBugSamppr}
\alias{collateBugSamppr}
\title{Collate and format sample-pair data for applying macroinvertebrate model predictions}
\usage{
collateBugSamppr(sampprTable, AttImp_L9 = "current (2006)",
  AttForest_L35W1000 = "current (2006)", meanAnnQ_mm = "current (2006)",
  mnAnnAirTm_deg = "long-term", SRI_48mth_weighted = "sample date")
}
\arguments{
\item{sampprTable}{Table containing at least the fields \code{subc}, \code{nspring},\code{nriff},
and either \code{process} or \code{processN}. If the table also contains a samppr identification column,
it should be called 'samppr'}

\item{AttImp_L9}{either the default value or a vector of new values for all subcs, or a single value if
all sampprs are to have the same value}

\item{AttForest_L35W1000}{either the default value or a vector of new values for all subcs, or a single value if
all sampprs are to have the same value}

\item{meanAnnQ_mm}{either the default value or a vector of new values for all subcs}

\item{mnAnnAirTm_deg}{either the default values, or a single number, added to
mwstreams$mnAnnAirTm_deg (e.g. 1 = 1 degrees warmer)}

\item{SRI_48mth_weighted}{if default, a "date" field must be supplied, or if a vector
of length dim(sampprTable)[1], then those values are used}
}
\value{
The original \code{sampprTable} with additional columns required for model predictions
}
\description{
Collate and format sample-pair data for applying macroinvertebrate model predictions
}
\details{
If arguments \code{AttImp_L9}, \code{AttForest_L35W1000}, \code{meanAnnQ_mm}, \code{mnAnnAirTm_deg},
are default values, then 2006 values are drawn from the mwstreams table, matched by \code{subc}.
If \code{SRI_48mth_weighted} is the default, values are drawn from \code{sri48moW}, matched by  \code{subc}
and  \code{date}.  Alternatively, for these arguments, vectors of length dim(sampprTable)[1]
can be supplied to represent scenarios other than 2006 conditions.  If supplying such data, care should be
taken not to set values outside the experience of the model (see Walsh and Webb 2014).  For instance, the
maximum air temperature value in the experience of the models is 15.3, and setting mnAnnAirTm_deg to 2 (i.e.
adding 2 degrees to the long-term average) raises the mean air temperature to >15.3 for many lowland reaches.
}
\examples{
sampprs <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package="melbstreambiota", mustWork = TRUE),
                                        sheet = 1))
#When importing data, replace the system.file() statement with "path/excelFile.xls"
sampprs1 <- collateBugSamppr(sampprs)
}
