% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_functions.R
\name{collateObsTable}
\alias{collateObsTable}
\title{Collate and format observed table for LUMaR calculation from macroinvertebrate data}
\usage{
collateObsTable(bugData, sampprTable)
}
\arguments{
\item{bugData}{Table with either A) 2 columns (\code{samppr}, \code{bugcode}), or A) one row per sample pair,
with the first column \code{samppr} and the remaining column headings \code{bugcode}s}

\item{sampprTable}{Table containing at least the fields \code{subc}, \code{nspring},\code{nriff},
and either \code{process} or \code{processN} (see \code{\link{collateBugSamppr}})}
}
\value{
A data.frame with 56 columns, the first samppr, matching sampprs$samppr, and the rest with names
equal to \code{taxon.classes$fam}. Presence are recorded as 1s and absences as 0s.
}
\description{
Collates macroinvertebrate data (in two possible formats) into an observed family-by-sample-pair
matrix for calculation of LUMar (and other indices).  The function also compares the macroinvertebrate
data to the sample-pair table to ensure consistency.
}
\details{
The vector of unique \code{samppr}s in bugData must completely match \code{sampprTable$samppr}. If bugData is in form A,
it is assumed that any listed family is present in the sample-pair (i.e. there are no taxa listed with abundance  = 0).
If there are duplicate entries of a family in a sample-pair, it is recorded as a single presence.  If bugData is in
form B, any family other than the 59 \code{taxon.classes$fam}s are removed, and any of the 59 that are missing are inserted
with abundance zero.
All bugcodes should be consistent with taxon.classes. See \url{http://www.mdfrc.org.au/bugguide/resources/AUSRIVAS_Taxacodes.pdf}
for a full list including taxa other than those in taxon.classes.  Note that the convention used in this package is to
omit the trailing 9999 from each code. If bugcodes in bugData include trailing 4 digits, these are removed
(this effectively combines any lower-taxonomic-level taxa to family-level.)
Taxonomic conventions used in the models are standard family-level with the following exceptions:
Chironomidae are identified to sub-family. The tribes of sub-family Chironominae ("QDAG","QDAH","QDAI") are combined to "QDAJ",
'Old' odonate taxonomy is used: new families split from Corduliidae ("QO23", "QO26", "QO28", "QO27", "QO30")
are consolidated to "QO16", and two families formerly considered Aeshnidae ("QO21", "QO12") are combined as "QO12".
If bugData contains any of these family, this function converts them as above.
}
\examples{
#Data in Form A
sampprTable <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                    package="melbstreambiota", mustWork = TRUE),
                                       sheet = 1))
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package="melbstreambiota", mustWork = TRUE),
                                        sheet = 2))
#When importing data, replace the system.file() statement with "path/excelFile.xls"
obsTable1 <- collateObsTable(bugData, sampprTable)
#Data in Form B
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package = "melbstreambiota", mustWork = TRUE),
                                        sheet = 3))
obsTable2 <- collateObsTable(bugData, sampprTable)
#Check that data imported in both forms are equql
sum(obsTable1[-1] - obsTable2[-1])  # = 0
}
