% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation_functions.R
\name{collateSampleFP}
\alias{collateSampleFP}
\title{Collate and format sample data for applying fish or platypus model predictions}
\usage{
collateSampleFP(
  sampleTable,
  FishOrPlatypus = "fish",
  AttImp_L9 = "current (2006)",
  AttForest_L35W1000 = "current (2006)",
  meanAnnQ_mm = "current (2006)",
  mnAnnAirTm_deg = "long-term",
  SRI_48mth_weighted = "sample date",
  barriersFromYear = TRUE,
  PartBarriersDS = NA,
  FullBarriersDS = NA
)
}
\arguments{
\item{sampleTable}{Table containing at least the field \code{subc}}

\item{FishOrPlatypus}{Predictor variables to be assembled for "fish" (the default) or "platypus"}

\item{AttImp_L9}{either the default value or a vector of new values for all subcs, or a single value if
all sampprs are to have the same value. If FishOrPlatypus = "platypus", AttImpMin4k_L9 is substituted, and the
field AttImpMin4k_L9.}

\item{AttForest_L35W1000}{either the default value or a vector of new values for all subcs, or a single value if
all sampprs are to have the same value}

\item{meanAnnQ_mm}{either the default value or a vector of new values for all subcs}

\item{mnAnnAirTm_deg}{either the default values, or a single number, added to
mwstreams$mnAnnAirTm_deg (e.g. 1 = 1 degrees warmer)}

\item{SRI_48mth_weighted}{if default, a "date" field must be supplied, or if a vector
of length dim(sampleTable)[1], then those values are used}

\item{barriersFromYear, }{if TRUE and FishOrPlatypus = "fish", a "date" field must be supplied,
if FALSE values must be provided for nPartBarriersDS, nFullBarriersDS}

\item{PartBarriersDS}{Only use if barriersFromYear is FALSE and FishOrPlatypus = "fish".
Must be a single numeric value (i.e. a number of barriers) to be applied ot all samples, or
a vector of values of length(SampleTable$subc) to be applied to each sample.}

\item{FullBarriersDS}{As for nPartBarriersDS}
}
\value{
The original \code{sampleTable} with additional columns required for model predictions
}
\description{
Collate and format sample data for applying fish or platypus model predictions
}
\details{
If arguments \code{AttImp_L9}, \code{AttForest_L35W1000}, \code{meanAnnQ_mm}, \code{mnAnnAirTm_deg},
are default values, then 2006 values are drawn from the mwstreams table, matched by \code{subc}.
If \code{SRI_48mth_weighted} is the default, values are drawn from \code{sri48moW}, matched by  \code{subc}
and\code{date}.  If \code{barriers} is the default, then "nPartBarriersDS" and
"nFullBarriersDS" are drawn from the barriers fields in mwstreams that match the year in which
each sample was collected (as determined by  \code{date}.  Alternatively, for these arguments, vectors of length dim(sampleTable)[1]
can be supplied to represent scenarios other than 2006 conditions.  If supplying such data, care should be
taken not to set values outside the experience of the model (see Walsh and Webb 2014).  For instance, the
maximum air temperature value in the experience of the models is 15.3, and setting mnAnnAirTm_deg to 2 (i.e.
adding 2 degrees to the long-term average) raises the mean air temperature to >15.3 for many lowland reaches.
}
\examples{
# A sample table that would be used for fish samples taken on the same day as the example
# macroinvertebrate samples
sampprs <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                         package="melbstreambiota", mustWork = TRUE),
                                        sheet = 1))
#When importing data, replace the system.file() statement with "path/excelFile.xls"
sampprs1 <- collateSampleFP(sampprs)
#A hypothetical sample from a subc in the mwstreams table (average antecedent
#flow, all fish barriers removed)
exampleFishSamp <- collateSampleFP(sampleTable = mwstreams[1,],
                                   FishOrPlatypus = "fish",
                                   SRI_48mth_weighted = 1,
                                   barriersFromYear = FALSE,
                                   PartBarriersDS = 0, FullBarriersDS = 0)
exampleFishSamp
}
