% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculation_functions.R
\name{lumar}
\alias{lumar}
\title{Calculate LUMaR index and associated statistics from observed and expected tables}
\usage{
lumar(obs.table, exp.table)
}
\arguments{
\item{obs.table}{a data.frame with first column samppr (or row.names = samppr), and 59 columns with names = \code{taxon.classes$fam}.
Data for each samppr are 1s for observed presences and 0 for absences}

\item{exp.table}{a data.frame with the same structure as obs.table, but filled with probabilities of occurrence
as predicted by the 59 distribution models described by Walsh (in prep) under no human impact}
}
\value{
A data frame with the same number of rows as the input tables and 14 columns: the LUMaR index and its components
#' \describe{
  \item{samppr}{\code{samppr} as in input tables}
  \item{lumar}{LUMaR index}
  \item{nSensFams}{Number of sensitive families (as defined by Walsh 2017) in each sample}
  \item{obs.wt.sens}{}
  \item{obs.wt.unexp}{}
  \item{obs.wt.A}{}
  \item{obs.wt.B}{}
  \item{obs.wt.C}{}
  \item{obs.wt.D}{}
  \item{obs.wt.weedy}{}
  \item{unexp.wt.B}{}
  \item{unexp.wt.D}{}
  \item{unexp.wt.weedy}{}
  \item{unexp.wt.invas}{}
  \item{exp.wt}{}
  }
}
\description{
Calculate LUMaR index and associated statistics from observed and expected tables
 formatted to include the 59 families in the same order as taxon.classes$fam
}
\examples{
sampprs <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package = "melbstreambiota", mustWork = TRUE),
                                        sheet = 1))
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package = "melbstreambiota", mustWork = TRUE),
                                        sheet = 2))
sampprs <- collateBugSamppr(sampprs)
expTable <- predCurrNHI59(sampprs)$predNHI
obsTable <- collateObsTable(bugData, sampprs)
lumar(obsTable, expTable)
}
\seealso{
\code{\link{collateBugSamppr}}, \code{\link{predCurrNHI59}}, \code{\link{collateObsTable}}
}
