% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mwCoastMap}
\alias{mwCoastMap}
\title{Map of the Melbourne region coastline}
\format{
A SpatialLinesDataDataFrame of 37 rows (polylines) and 1 columns
\describe{
  \item{ID}{unique ID}
}
}
\source{
\url{http://land.vic.gov.au}
}
\usage{
mwCoastMap
}
\description{
A SpatialLinesDataDataFrame of the coastline
}
\keyword{datasets}
