% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_calculation_functions.R
\name{oestats8}
\alias{oestats8}
\title{All index variants considered by Walsh (in prep)}
\usage{
oestats8(obs.table, exp.table, general.threshold = NA)
}
\arguments{
\item{obs.table}{a data.frame with first column samppr (or row.names = samppr), and 59 columns with names = \code{taxon.classes$fam}.
Data for each samppr are 1s for observed presences and 0 for absences}

\item{exp.table}{a data.frame with the same structure as obs.table, but filled with probabilities of occurrence
as predicted by the 59 distribution models described by Walsh (in prep) under no human impact}

\item{general.threshold}{if not NA, a single proportion signifiying the presence-absence threshold to be applied to all families}
}
\value{
A data frame with the same number of rows as the input tables and 28 columns
#' \describe{
  \item{samppr}{\code{samppr} as in input tables}
  \item{OE50}{}
  \item{wOE50}{}
  \item{OE25}{}
  \item{wOE25}{}
  \item{OE.prev}{}
  \item{wOE.prev}{}
  \item{OE.pa.prev}{}
  \item{wOE.pa.prev}{}
  \item{wOE}{}
  \item{wOE.signal}{}
  \item{wOE.signal1}{}
  \item{wOE.signal.prev}{}
  \item{OE.signal.pa.prev}{}
  \item{simOE}{}
  \item{simOE.prev}{}
  \item{wOE.sensif}{}
  \item{wOE.sensif1}{}
  \item{wOE.sensif.prev}{}
  \item{OE.sensif.pa.prev}{}
  \item{wOE.sensif.minweed.prev}{}
  \item{wOE.sensif.minweed.prevA}{}
  \item{wOE.sensif.minweed.prevB}{}
  \item{wOE.sensif.minweed.prev1}{}
  \item{wOE.sensif.minweed.prev1A}{}
  \item{wOE.sensif.minweed.prev1B}{}
  \item{wOE.signal.minweed.prev}{}
  \item{wOE.signal.minweed.prev1}{}
  }
}
\description{
Calculate all OE indices, and others, compared to LUMaR by Walsh (in prep). Note that SIGNAL indices n
  need to be calculated using the full list of families collected in samples, not just the 59 families used in this
  function
}
\examples{
sampprs <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package="melbstreambiota", mustWork = TRUE),
                                        sheet = 1))
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                     package="melbstreambiota", mustWork = TRUE),
                                        sheet = 2))
sampprs <- collateBugSamppr(sampprs)
expTable <- predCurrNHI59(sampprs)$predNHI
obsTable <-collateObsTable(bugData, sampprs)
oestats(obsTable, expTable)
}
\seealso{
\code{\link{collateBugSamppr}}, \code{\link{predCurrNHI59}}, \code{\link{collateObsTable}}
}
