% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{plotMWstreams1col}
\alias{plotMWstreams1col}
\title{Plot a map of Melbourne's streams (or a subset) by a single colour}
\usage{
plotMWstreams1col(line_col, subcSubset = NA, ...)
}
\arguments{
\item{line_col}{a single colour}

\item{subcSubset}{if NA the whole stream network is mapped.
If a vector of subcs (a subset of mwstreams$subc) is provided, only that subset will be mapped.}

\item{...}{other arguments to be passed to \code{classInt::classIntervals} or \code{sp::plot}}
}
\value{
A plot of the streams of Melbourne
}
\description{
Plot a map of Melbourne's streams (or a subset) by a single colour
}
\details{
A more flexible option for \code{plotMWstreamsByVar}, when var is a single value
}
