% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{plotMWstreamsByVar}
\alias{plotMWstreamsByVar}
\title{Plot a map of Melbourne's streams colour-coded by a variable}
\usage{
plotMWstreamsByVar(
  var,
  nbreaks = 5,
  style = "quantile",
  varName = "Variable",
  palette = "Spectral",
  rev = FALSE,
  legend = TRUE,
  legend.cex = 1.5,
  fixedBreaks = NA,
  subcSubset = NA,
  ...
)
}
\arguments{
\item{var}{A vector (length 8246) of variable values for each subc in mwstreams (the function assumes that the vector is in the same order as mwstreams)}

\item{nbreaks}{number of classes to be plotted: equals n argument in classInt}

\item{style}{chosen style as in \code{classIntervals}: one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", or "jenks}

\item{varName}{character string for legend title}

\item{palette}{an RcolorBrewer palette (see \code{RColorBrewer::brewer.pal})}

\item{rev}{inverts palette if TRUE}

\item{legend}{plots legend if TRUE}

\item{legend.cex}{character expansion factor for legend}

\item{fixedBreaks}{NA unless style = "fixed", and then must equal a vector of breaks with length nbreaks + 1}

\item{subcSubset}{if NA the whole stream network is mapped.
If a vector of subcs (a subset of mwstreams$subc) is provided, only that subset will be mapped.}

\item{...}{other arguments to be passed to \code{classInt::classIntervals} or \code{sp::plot}}
}
\value{
A plot of Melbourne's streams.
}
\description{
Plot a map of Melbourne's streams colour-coded by a variable
}
\details{
The "fixed" style permits a \code{classIntervals} object to be specified with given breaks,
set in the fixedBreaks argument; the length of fixedBreaks should be nbreaks+1; this style can be
used to insert rounded break values.
The "jenks" style is not implemented.
}
\examples{
plotMWstreamsByVar(melbstreambiota::mwstreams$AttForest_L35W1000, style = "fixed",
                   fixedBreaks = seq(0,1,0.2), varName = "Attenuated Forest Cover",
                   legend.cex = 0.75)
}
\seealso{
\code{\link[classInt]{classIntervals}}, \code{\link[RColorBrewer]{brewer.pal}}
}
