% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumar_diagnostic_functions.R
\name{plotSensFamDiagnostics}
\alias{plotSensFamDiagnostics}
\title{Which families drive differences in LUMaR between groups of sample-pairs?}
\usage{
plotSensFamDiagnostics(
  lumarObject,
  covariate,
  covariateName = "Covariate",
  sensLinEffects,
  famLinEffects
)
}
\arguments{
\item{lumarObject}{a data.frame produced by the function \code{lumar}}

\item{covariate}{a vector length = number of rows in lumarObject (that explains some of the variation in LUMaR scores)}

\item{covariateName}{a character vector for labelling the output plot}

\item{sensLinEffects}{...}

\item{famLinEffects}{...}
}
\value{
A plot identifying variation of observed LUMaR score from predicted values under 2006
conditions and under no human impact, as well as the most influential families driving any differences. If returnTable
- TRUE a list of 2 data.frames is also returned:
\describe{
  \item{InfluencesFromRef}{...}
  \item{InfluencesFromPred}{...}
  }
}
\description{
Diagnostic plots to identify the most influential families influencing differences between groups of sample pairs,
a trend among sample-pairs explained by a covariate
}
\examples{
sampprs <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                package = "melbstreambiota", mustWork = TRUE),
                                    sheet = 1))
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                package = "melbstreambiota", mustWork = TRUE),
                                    sheet = 2))
sampprs <- collateBugSamppr(sampprs)
expTables <- predCurrNHI59(sampprs)
obsTable <- collateObsTable(bugData, sampprs)
exampleLumar <- lumar(obsTable, expTables$predNHI)
stream <- substr(exampleLumar$samppr, 1, 3)
# What families explain differences in LUMaR between the two streams (BRS and OLN)
famStreamEffects <- famCovariateEffects(obsTable, expTables$predNHI, stream)
sensStreamEffects <- sensGpCovariateEffects(exampleLumar, stream)
plotSensFamDiagnostics(exampleLumar, stream, covariateName = "stream",
                       sensLinEffects = sensStreamEffects,
                       famLinEffects = famStreamEffects)
}
