% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prediction_functions.R
\name{predCurrNHI59}
\alias{predCurrNHI59}
\title{Compile predictions of occurrence for all 59 macroinvertebrate families in a set of sample-pairs}
\usage{
predCurrNHI59(sampprPreds)
}
\arguments{
\item{sampprPreds}{a data.frame of sample-pair codes (samppr) each with
associated environmental predictors. If there is no field called samppr, then the field subc if used as samppr.}
}
\value{
A list of two dataframes with number of rows = that of \code{sampprPred},
(each row corresponding to the equivalent row in \code{sampprPred}) and 59
columns, corresponding to the 59 families listed in \code{taxon.classes}
\describe{
  \item{predCurr}{predicted probabilities of occurrence under conditions specified in \code{sampprPred}}
  \item{predNHI}{predicted probabilities of occurrence under no human impact
  (AttImp_L9 set to zero, AttForest_L35W1000 set to 1)}
  }
}
\description{
Compile predictions of occurrence for all 59 macroinvertebrate families in a set of sample-pairs
}
\details{
The sampprPreds table must contain the 10 predictor variables used in the
models: nspring,nriff,processN,AttImp_L9, AttForest_L35W1000,
CatchmentArea_km2_InclDams, meanAnnQ_mm,mnAnnAirTm_deg,
CatIgneous, and SRI_48mth_weighted. The function \code{collateBugSamppr} can
be used to compile such a table.  Use \code{bugModPred1} to make equivalent predictions for a single taxon.
}
\seealso{
\code{\link{collateBugSamppr}}, \code{\link{bugModPred1}}, \code{\link{taxon.classes}}
}
