% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumar_diagnostic_functions.R
\name{sampprDiagnostic}
\alias{sampprDiagnostic}
\title{Interpretation of an observed LUMaR score at an individual site}
\usage{
sampprDiagnostic(
  samppri,
  obsTable,
  currentPredTable,
  NHIPredTable,
  returnTable = FALSE
)
}
\arguments{
\item{samppri}{a data.frame with a single row, with fields samppr, and subc
(a subc value drawn from mwstreams).}

\item{obsTable}{a data.frame with first column samppr
(or row.names = samppri) (the data can contain multiple rows,
but one row must contain the value of samppri$samppr) and 59 columnns with
names matching taxon.classes$fam.  ' Values = 1 for presences and 0 for absences.
under no human impact}

\item{currentPredTable}{A data.frame the same structure as obs.table, but filled with probabilities of
occurrence under current conditions as predicted by the 59 distribution models described by Walsh (2017).}

\item{NHIPredTable}{A data.frame the same structure as obs.table, but filled with probabilities of
occurrence under no human impact as predicted by the 59 distribution models described by Walsh (2017)}

\item{returnTable}{logical, if TRUE, returns 2 tables in addition to plotting results}
}
\value{
A plot identifying variation of observed LUMaR score from predicted values under 2006
conditions and under no human impact, as well as the most influential families driving any differences. If returnTable
- TRUE a list of 2 data.frames is also returned:
\describe{
  \item{InfluencesFromRef}{...}
  \item{InfluencesFromPred}{...}
  }
}
\description{
Interpretation of an observed LUMaR score at an individual site
}
\examples{
sampprs <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                package = "melbstreambiota", mustWork = TRUE),
                                    sheet = 1))
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                package = "melbstreambiota", mustWork = TRUE),
                                    sheet = 2))
sampprs <- collateBugSamppr(sampprs)
expTable <- predCurrNHI59(sampprs)
obsTable <- collateObsTable(bugData, sampprs)
sampprDiagnostic(samppri = sampprs[8,],
                 obsTable = obsTable, currentPredTable = expTable$predCurr,
                 NHIPredTable = expTable$predNHI)
}
