% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumar_diagnostic_functions.R
\name{sensGpCovariateEffects}
\alias{sensGpCovariateEffects}
\title{Find the relative contribution of taxon-sensitivity groups to differences in LUMaR associated with a covariate}
\usage{
sensGpCovariateEffects(lumarObject, covariate)
}
\arguments{
\item{lumarObject}{a data.frame as produced by the \code{lumar} function}

\item{covariate}{a vector of covariate values corresponding to samppr used as a contrast in LUMaR scores}
}
\value{
A data frame with rows corresponding to the families found or expected in the sampprs (<= 59) and 6 columns:
\describe{
  \item{fam}{the \code{bugcode}s matched by \code{taxon.classes$fam}}
  \item{unexp}{0 = families whose presence contributed to differences in LUMaR by their sensitivity grade;
  1 = families whose presence contributed by their unexpected grade.}
  \item{coef}{the coefficient of a beta regression of the weighted occurrences of each families against the covariate}
  \item{lwr}{the 10th percentile confidence interval of coef}
  \item{upr}{the 90th percentile confidence interval of coef}
  \item{sig}{'significance': TRUE if the range between lwr and upr includes zero}
  }
}
\description{
Find the relative contribution of taxon-sensitivity groups to differences in LUMaR associated with a covariate
}
\examples{
sampprs <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                package = "melbstreambiota", mustWork = TRUE),
                                    sheet = 1))
bugData <- as.data.frame(readxl::read_excel(system.file("extdata","exampleData.xlsx",
                                                package = "melbstreambiota", mustWork = TRUE),
                                    sheet = 2))
sampprs <- collateBugSamppr(sampprs)
expTables <- predCurrNHI59(sampprs)
obsTable <- collateObsTable(bugData, sampprs)
exampleLumar <- lumar(obsTable, expTables$predNHI)
stream <- substr(exampleLumar$samppr,3,5)
#What families explain differences in LUMaR between the two streams (BRS and OLN)
sensGpCovariateEffects(exampleLumar, stream)
}
\seealso{
\code{\link{collateBugSamppr}}, \code{\link{predCurrNHI59}}, \code{\link{collateObsTable}}
}
