% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sri48moW}
\alias{sri48moW}
\title{Monthly 48-month linearly-weighted antecedent discharge since 1986 for all mwstream subcs}
\format{
a data frame 3 columns
\describe{
  \item{date}{\code{POSIXct} date, time zone "UTC"}
  \item{subc}{subcatchment ID equivalent to \code{mwstreams$subc}}
  \item{SRI_48mth_weighted}{48-month linearly-weighted antecedent discharge as a proportion of mean annual discharge}
  }
  @details The general method for \code{SRI_48mth_weighted} was described by
  Walsh and Webb 2014. These data use mean monthly estimates of catchment
  runoff derived from daily runoff estimates Australian Landscape Water
  Balance (AWRA-L) model (Australian Bureau of Meteorology).  The \code{date}
  field equals the first of each month from March 1981 to the end of the most
  recent full year. This file is stored at https://osf.io/mcxrq/ (and
  downloaded when the package is installed). It is updated early each year to
  include the previous year's data. To update the cached version of this file
  run the function \code{update_sri()}.
}
\source{
Walsh, C. J. (2023) LUMaR: a sensitive macroinvertebrate
index of stream condition combining observed:expected ratios and sensitivity
weightings. OSF preprints. https://osf.io/392kv
}
\usage{
sri48moW
}
\description{
A data frame data of subcs each with monthly 48-month linearly-weighted
antecedent discharge since (1986-2014 [some to 2016]) for all mwstream subcs,
calculated as described by Walsh and Webb (2014, 2016)
}
\keyword{datasets}
