% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streamsubcs.allus2017}
\alias{streamsubcs.allus2017}
\title{A list of subcatchments each with a list of all upstream subcatchments that contain stream reaches in mwstreams}
\format{
a list of length 8257
}
\usage{
streamsubcs.allus2017
}
\description{
A list of subcatchments each with a list of all upstream subcatchments that contain stream reaches in mwstreams
}
\keyword{datasets}
