% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon.classes}
\alias{taxon.classes}
\title{Details of the 59 modelled families}
\format{
a data frame of 59 rows and 10 columns
\describe{
  \item{fam}{bugcode, consistent with EPA Victoria and AUSRIVAS conventions}
  \item{family}{family name}
  \item{devexp}{proportion explained deviance explained by the model}
  \item{pa.threshold}{presence-absence threshold used in the LUMaR index}
  \item{sens.gp}{sensitivity group used in the LUMaR index}
  \item{SIGNAL}{SIGNAL grade (Chessman 1995)}
  \item{SIGNAL2}{SIGNAL2 grade (Chessman 2003)}
  \item{unexp.grade}{sensitivity grade used in the LUMaR index}
  \item{sens.grade}{unexpected grade used in the LUMaR index}
}
}
\source{
Walsh, C. J. (in preparation) LUMaR: a sensitive macroinvertebrate
index of stream condition combining observed:expected ratios and sensitivity
weightings.
}
\usage{
taxon.classes
}
\description{
A data frame listing the 59 families that
are modelled in \code{bestModelsBugfams}, with relevant statistics
}
\keyword{datasets}
