% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vertSpp}
\alias{vertSpp}
\title{Fish species and platypus groups modelled}
\format{
a data frame of 25 rows and 4 columns
\describe{
  \item{vertcode}{the code used to identify models in the package}
  \item{species}{}
  \item{authority}{}
  \item{Commmon name}{}
}
}
\usage{
vertSpp
}
\description{
A data frame listing 23 fish species and associated codes (and two platypus groups) used in models
}
\keyword{datasets}
