% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sri48moW}
\alias{sri48moW}
\title{Monthly 48-month linearly-weighted antecedent discharge since 1986 for all mwstream subcs}
\format{a data frame of 2806548 rows and 3 columns
\describe{
  \item{date}{\code{POSIXct} date, time zone "UTC"}
  \item{subc}{subcatchment ID equivalent to \code{mwstreams$subc}}
  \item{SRI_48mth_weighted}{48-month linearly-weighted antecedent discharge as a proportion of mean annual discharge}
  }
  @details The general method for \code{SRI_48mth_weighted} was described by Walsh and Webb 2014.
  These data are the variant, calculated from the Australian Water Availability Project data
  (AWAP; Raupach et al. 2009) as described in Walsh (in preparation).  The \code{date} field equals
  the first of each month from Jan 1986 to Feb 2014 for most subcs, and to July 2016 for some (Main Creek,
  French Island, Bass, Lang Lang, Yallock, and Bunyip catchments).}
\source{
Walsh, C. J. (in preparation) LUMaR: a sensitive macroinvertebrate
index of stream condition combining observed:expected ratios and sensitivity
weightings.
}
\usage{
sri48moW
}
\description{
A data frame data of subcs each with monthly 48-month linearly-weighted
antecedent discharge since (1986-2014 [some to 2016]) for all mwstream subcs,
calculated as described by Walsh and Webb (2014, 2016)
}
\keyword{datasets}
